requirements.txt
Flask==3.1.2
Werkzeug==3.1.3
Jinja2==3.1.4
MarkupSafe==2.1.5
itsdangerous==2.2.0
click==8.1.8
pandas==2.2.2
numpy==1.26.4
python-dateutil==2.9.0.post0
pytz==2024.1
tzdata==2023.3
Pillow==10.3.0
opencv-python==4.11.0.86
matplotlib==3.9.2
pyparsing==3.1.2
cycler==0.12.1
fonttools==4.51.0
kiwisolver==1.4.5
contourpy==1.2.0
torch==2.5.1
torchvision==0.20.1